Try to remake protocol. New's protocol structure: - magicByte (0x53) - 1 byte, - address - 1 byte, - length of data+crc16 - 1 byte, - crc8 - 1 byte - control sum of header, - command - 1 byte, - pktNumber - 1 byte, - data, - crc16 - control sum of data. This program need to modify: - crc16 not only for data, but for all packet, - length field have to consist all packet.

+++ src/ConfigCommand.java
@@ -2,7 +2,7 @@
  * Created by apu on 04.11.2016.
  */
 public class ConfigCommand {
-    public static final int COMMAND_STATUS  = 0;
-    public static final int COMMAND_READ    = 1;
-    public static final int COMMAND_WRITE   = 2;
+    public static final byte COMMAND_STATUS  = 0;
+    public static final byte COMMAND_READ    = 1;
+    public static final byte COMMAND_WRITE   = 2;
 }
